AbstractElasticIndex: fix index name of delete request and add coverage
Previously the deletion of the index failed because of the improper
index name. This change fixes the index name of the delete request so
that the objects (i.e., groups, accounts or changes) that exist in the
index can be deleted.

This change also adds the test to delete a change from the index in
order to test this fix and add coverage for the getDeleteActions method
in ElasticChangeIndex.

Bug: Issue 9040
Change-Id: I891bd4dc286f7bf0cfeb95de1c952d91ae631ba4

Former-commit-id: bb412404dc125260e5883883b95022ac3751a6c0