Only permit registering username when allowed
The gr-account-details form used in settings houses logic for when to
display an editable username. In particular, the username is only
editable when the server config lists "USER_NAME" as an "editable
account field" and a username hasn't already been set.

However, this logic was not used in the gr-registration-dialog form.
Rather, an input for username was always shown, whether or not it was
editable. As a result, if a user tried to fill out a form when the
username could not be saved, it would always result in an error.

Use the same logic as gr-account-details in gr-registration-dialog for
deciding whether the username should be editable.

Bug: Issue 8167
Change-Id: Ia614b5996f0bcfe7fdf785efcebbbde27919e7d3

Former-commit-id: 9a1e03f9328421577a9933c22807377a4ba7f94c