Add keyword type to index type system
Elasticsearch 5.0 introduced different ways to search strings: [1],
either search whole values that we often refer to as keyword search,
or individual tokens, that we usually refer to as full-text search.

To reflect this change and support such fields in gerrit with ES 5 and
higher, add keyword type to gerrit index type system.

[1] https://www.elastic.co/blog/strings-are-dead-long-live-strings
Change-Id: I25bd157cd638695f5a3b533463008fcf7da9ac8d

Former-commit-id: 506125e71a786f31573e6ba1b1983a4c48cc1dbc