GroupField: Change UUID fields' type to KEYWORD
This is a preparatory change for adding support for Elasticsearch
version 5, where handling of keyword types is changed.

For the Lucene index, handle KEYWORD the same way as EXACT.

For Elasticsearch, also handle both KEYWORD and EXACT in the same way
since at this point we don't yet support Elasticsearch 5.

Change-Id: I545c3ce85b44f6f640c51afd22ba3c663ff6ba80

Former-commit-id: 09fdc3eaadd352b628acfc95c5b74df305b9dbf9