AbstractElasticIndex: Rename getActions to getDeleteActions
The getActions method is only ever called from the delete(K) method to
add the actions for deleting an item. It's only needed because in the
change index implementation it is necessary to add multiple actions.

Rename it to getDeleteActions which is more aligned with its purpose.

Also clean up the parameter names in these methods. Instead of naming
them all 'c', which is leftover from when we only had the change index,
rename them to 'id' in the abstract class, and to 'a' and 'g' in the
account and group index implementations.

Change-Id: Ib1dde44f12147a5396b3296666aafefcd9e4846b

Former-commit-id: c46fb11fd1adcf5e2f8c1f4a0fe2916dbc3193f1