Use CacheModule for the external ID cache again
In c1be869e we stopped using CacheModule for this cache, because we
thought the value should not be configurable. However, there are a few
advantages to using CacheModule that we would like to keep:
 * It allows us to expose metrics from the caches returned by
   DefaultMemoryCacheFactory, and also get these in the ExternalIdCache.
 * It allows us to explore persisting the ExternalIdCache as an optional
   optimization, using the normal cache persistence path.

Revert to using CacheModule, and simply document that we do not
recommend changing the maximum size.

This reverts commit 6d717abcae221f9cb681ccdc9a033707c3883c8c [formerly c1be869e1d38c811d28fa201006367daeb73b071].

Change-Id: Ie3505a6bab4d31b67384feaebcab20388a7ae1b4

Former-commit-id: 6f7fa0a0da3cf40a7b38ab6eacd92e2441f5f0c4