Support expiring from in-memory caches after last access
Use the existing CacheBuilder#expireAfterAccess method for in-memory
caches.

This configuration is not respected for persistent H2 caches. There is
already an "accessed" column, so in theory we could compare the current
time to the value in this column. However, this has the problem that the
column is not updated on every lookup against the outer memory cache
which wraps the H2 cache; it's only updated when the value isn't in the
memory cache and needs to be read into the memory cache from storage. As
a result, the most frequently accessed elements in the in-memory cache
would be the first elements to expire from the persistent cache.

Rather than allowing this perverse behavior, just ignore
expireAfterAccess in the H2 implementation, and document it properly.
Elements may still get pruned from the cache when it exceeds the
configured size. (The pruning code actually uses the accessed column
and thus may be subject to the perverse behavior we're trying to avoid,
but fixing it is beyond the scope of this change.)

It would be theoretically possible to update the accessed timestamp in
the H2 database on every read from the memory cache, but cache reads are
extremely frequent and this would likely put too much write load on the
H2 database.

Change-Id: I881df14bcabde1376145eed815f69d7493751574

Former-commit-id: 08bac5cac3343bcc2a8d4dc4115fe30cd616721b