Elasticsearch: fix changes index type for bulk calls
Previously, when a bulk call was made to index or delete some changes,
the index type passed to the request was "changes", which does not
exist in the elasticsearch index. This change ensures valid index type
for bulk calls for changes, namely "open_changes" and/or
"closed_changes" for Elasticsearch V2 and V5, and "_doc" for
Elasticsearch V6.

The index types are valid for indexing or deleting requests for accounts
and groups ("accounts" and "groups" respectively).

Change-Id: I6f06281e32a54db9c8ab7a8a5724fa3de78d6a3e

Former-commit-id: 408b3f0adbb26620c2112efaa3a463fdfb0b70de