Set labels on revert via review REST API method
To satisfy issue 4675, revert actions may be followed by posting a
plugin-defined set of labels on the new revert change so that reverts
can be immediately submittable. gr-change-actions accomplished this by
using the same style of REST API interface call that is used for the
change actions themselves.

Because setting labels is actually a call to posting a review, switch
over to using the `saveChangeReview` method. This change is one step
along the way to moving gr-change-actions off of the
`getChangeURLAndSend` utility method.

Change-Id: I3b6b1f00a2164d4a6bba7beda5ee2dd13818f100

Former-commit-id: 5ed53c1a901ee5dc656ebdb6f8a0c671a82981b0