Bazel: Remove unused commons-lang3
After removing Elasticsearch and Jest REST client library, commons lang3
is not used any more in Gerrit core. Given that we have forgotten to
remove it and stil expose it in plugin API only confused the plugins
that depend on it, because the build pass, but then the plugin cannot
work, because the required dependency is missing.

Change-Id: I7af2f9f6cb90e8e52d7ed2efaf4c3af85b47a9e2

Former-commit-id: 589af813ccca4c329bd4845852cd9d6d73d1d11d