CreateChange: Strip comment lines out of input commit message
Attempting to create a change with only a commit message subject where
the subject begins with a # fails due to an invalid (zeros) Change-Id.

The zero Change-Id is appended to the change because the JGit utility
method used to generate it strips the commit message of any comment
lines, which in this case results in an empty commit message, and then
the commit message does not have enough content.

Strip comment lines out of the input commit message before attempting
to use it to create the change. If it is empty after stripping the
comments, return 400 Bad Request.

Update the 'ChangeInput' documentation to mention that comment lines
will be stripped. Also change the description to reflect the fact that
the `subject` field is actually used for the entire commit message,
rather than only the subject.

Bug: Issue 9195
Change-Id: I5a2eed97aeb187801550f3a6b35e3dcfd2baf34e

Former-commit-id: 38cb32d7b739a147f0fa48d0796d2fee17250f60