Migrate NoteDb comments to JSON
In 2.15-rc2, new sites are created by default with NoteDb as the only
storage for changes, but comment notes were still written with the
legacy (non-JSON) format. We need to remove support for the legacy
format eventually, but we couldn't do a schema migration in the stable
branch. Do it in master instead.

For tests, we create a mixture of JSON and legacy formatted notes by
hand-crafting meta branch commits.

Change-Id: I377e22320a0013eaa36c80f36d30b6ec2c5afad8

Former-commit-id: 548d7222ba3888855eb1ce8e781139bf7da5c3dc