Remove unneeded checks from logDebug methods to check if fine logging is enabled
Calling the logVarargs method is expensive since it requires allocating
an object array, however in case of the logDebug methods the object
array was already allocated when calling the logDebug method, hence
checking inside of the logDebug methods if fine logging is enabled to
avoid calling the logVarargs method if it's not enabled doesn't help.

The custom logDebug methods are still needed to include request IDs,
submission IDs etc. In the long this should be done via Flogger tags and
the custom logDebug methods should be removed (but we don't have support
for Flogger tags yet).

Change-Id: Ia1a3048fa3d0ab2fbb801dc278d0d2d61e2c86dc
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: bf734b9d77ba5dcffd881ba6d13cf8af02abb723