RevisionIT: Add test for posting review comment on non-existing file
Test that when posting a review on a non-existing file, the correct
error is returned. Specifically, test that the message mentions the
same patch set number as used in the review request.

This tests the fix that was done in change I95a505d95, and without
that fix, the test fails.

Change-Id: Ib86d61a94494c8c7ab5dabc56e54006a3eec8a20

Former-commit-id: b002fe7ecab9e3ad63fb5104074577a910b95fb3