Move: Return the modified change in JSON response
The original change was returned as response, meaning the response's
content included the original branch name in the "branch" and "id"
fields.

Return the modified change instead.

Test plan:

Move a change using the REST API:

  curl -n -XPOST --header "Content-Type: application/json" -d@data.txt http://localhost:8080/a/changes/I4e38153d7c1960688f622e8b0e3bca037a8f0b6c/move

where data.txt contains:

 {
    "destination_branch" : "stable"
 }

and observe that the response contains the new branch name in the
"branch" and "id" fields.

Bug: Issue 9245
Change-Id: Ic2808fdf00f2232521dc02f07af88d8a1d0fb61f

Former-commit-id: f33b4e5aa9349178c1d83dd254a5b19c71cdc019