Fix editable-label losing focus
The editable label was setting focus on the dropdown immediately due to
it awaiting the wrong exit condition for _awaitOpen, resulting in a
race where focus was only sometimes set on the input.

The function was borrowed from gr-overlay, which has styles applied to
it directly. In the case of the editable label, we must wait for the
style to be set on the dropdown instead.

Bug: Issue 8980
Change-Id: I31439a2065159976a9aeda52efd366914cae617f

Former-commit-id: 06cf5833a378c5431fd5c3308793d5e05daaadd9