Make awaiting draft save feel even less blocking
When creating, updating or deleting draft comments, the diff containing
the draft does not need to be kept open for the underlying requests to
succeed. However, in a number of ways, the draft UI suggests that it
does. This results in the draft writing process feeling like a blocking
experience when it need not be.

Change the draft UI to feel less blocking:
- Instead of merely disabling the textarea when saving, the textarea is
  hidden and the presentation form of the message is shown immediately.
- Instead of disabling a draft as its being discarded, the draft
  disappears immediately.
- The `_savingMessage` is removed.

Bug: Issue 6970
Change-Id: Ia396f0f1780571def84be0d20f19c98bdb0b0cde

Former-commit-id: 8635c8effb9bf4d3c88efb0292ca084db5771565