Mark getChangeUrlAndSend private
The getChangeUrlAndSend method of the REST API interface is used
primarily as an internal utility function. However, it could not be
marked as private because it's used outside the interface by the change
actions handler.

Add a new, public method for executing actions and mark the
getChangeUrlAndSend method as private.

Change-Id: I4dab166b29f0f119b6f41789cfffabf0709ef0ea

Former-commit-id: f73adc47bae1ce3a4a89b2ccd776bbbde24ad0b3