Bazel: Disambiguate codemirror-minified repository
This repository is used twice: for GWT and for PG toolchains, under
different names:

* codemirror_minified for GWT
* codemirror-minified for PG

In I1e75690fe1 the external repository names were harmonized, but
renaming of codemirror_minified to codemirror-minified was undone
to avoid collision between the two external repositories.

Make them unique and rename codemirror_minified to
codemirror-minified-gwt to avoid collision.

Change-Id: I14c2f319d0ca451dea9424af9f5080d89dc704ad

Former-commit-id: d2f001b0c4feea2446d6a9c515e4469a54ec5e29