Remove commons-lang3 from plugin API
After removing Elasticsearch and Jest REST client library, commons lang3
is not used any more in Gerrit core and is only used for tests and not
included in release.war.

Given that we have forgotten to remove it and stil expose it in plugin
API only confused the plugins that depend on it, because the build pass,
but then the plugin cannot work, because the required dependency is
missing.

Similar change was already done on stable branches, but was erroneously
reverted during merge to master.

Bug: Issue 9253
Change-Id: Ie86b9078f5d6ba353ae2d044adefccb182a015ba

Former-commit-id: 02bc2e9beee8ab052945cb79d4a677db7222fb09