Elasticsearch: Adapt version discovery for version 6
In version 6 the /_aliases endpoint doesn't work with GET requests
and returns an error. Instead, we have to just make a GET against
the root URL with wildcard.

Extend the adapter to provide the correct URL.

Also make the discovery throw an exception if it failed to get the
versions, where previously it would silently return an empty list.

Bug: Issue 9267
Change-Id: I161347b4ff89d387c511b0dcc6323c93ddbe0b2b

Former-commit-id: f9c4632f11eb3c4c0af388d55cbe928df50cc37c