Elasticsearch: Add support for version 6.3.0
6.3.0 is now the latest version on the 6.x series. Update the version
detection to support this version, and add support for it in the query
adapter. There don't seem to be any breaking changes since 6.2.x, so
it's not necessary to add any specific handling other than modifying
some conditions to check for 6.3 in addition to 6.2.

Update the test container to support 6.3, using the -oss version of
6.3.0 from the elastic.co docker registry.

Update the V6 query tests to use 6.3 instead of 6.2. To test against
both 6.2 and 6.3 we would need to duplicate the account, group and
query test classes again resulting in 4 copies (V2, V5, V6.2, V6.3)
of each. If we need to explicitly test against 6.2 at any point in
the future we can manually change them back to use the 6.2 container.

Update the index and reindex tests to explicitly test 6.3 in addition
to 6.2, since those tests can be extended simply by adding another
@ConfigSuite.Config annotation.

Change-Id: I02afbac3e04958a13a6baae74e9d0de9cd3b644b

Former-commit-id: bc8d1c24f8fc29a728dfaafcc4fcb45eefed0474