AbstractIndexTests: Use correct command to index project
The correct command is "project", singular, not "projects" plural.

Fix this, and add assertions on the SSH command status which would
have caught this mistake earlier.

The test still needs further investigation, since it was passing
despite the project not actually being indexed (due to the wrong
command that was fixed here).

Change-Id: I24c1492fc74d5412e8a488d73456a9f541c45440

Former-commit-id: 44c5d542b58b8bdfd394db599facfc33fc1e9863