AbstractIndexTests: Add assertions on change index count
Add ChangeIndexedCounter based on the one added in ChangeIT in change
I66dda79bb. Use it to assert on the number of times the change gets
indexed during the tests.

For each test add the configuration to set index.autoReindexIfStale to
false, thus preventing the background reindexing of the change.

After these updates we can see that the project index test is failing.

Previously it was passing because:

- The SSH index command was wrong, but the status of the SSH command
  was not checked. (This was fixed in change I24c1492fc).

- The query returned the expected result because the project had been
  indexed in the background by the auto reindex.

The reason it now fails is because the index project command starts an
indexing task asynchronously and then immediately returns. The assertion
on the number of index calls is then done before the task completes.

The same test procedure passes for the index change command because it
indexes the change synchronously.

After the call to the index project command, repeatedly check the output
of the show-queue command until the "Index all changes of project" task
is not included. This indicates that the reindex was completed and the
test can proceed.

Change-Id: Iabd119b1eee495887c7f11e7fc6737ec4e39ca08

Former-commit-id: ba6125a747bfc832fca64b1795e7a9cd89473c7a