OutgoingEmail#getName{Email}For: Use helpers from Account
The getNameEmailFor and getNameFor methods do almost the same as the
helper methods Account.getNameEmail and Account.getName but with a
slight difference in the formatting. When the name/email is not set,
the getNameEmailFor and getNameFor methods format the name as:

  Anonymous Coward #123

where "123" is the account number. The Account methods format it as:

  Anonymous Coward (123)

Using the Account methods reduces duplicate code, and also makes the
formatting consistent with the rest of the application.

Change-Id: I9d55ec3f1eb09152548afb975ce29df14b0d21c3

Former-commit-id: ddf27f44c8798ef810386a138c5dda9ecdc7d42c