Update JGit to 5.0.0.201806131550-r
JGit 5.0.0 introduces some new features (including support for the
git protocol v2), performance improvements, and bug fixes.

It also includes a number of deprecations, some of which cause
warnings in the gerrit project. These are not fixed in this commit,
but will be fixed in follow-ups.

See the release notes [1] for full details of the release.

[1] https://projects.eclipse.org/projects/technology.jgit/releases/5.0.0

Change-Id: I62200941b34ebcfa58934be531a2486b1fef3d01
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>

Former-commit-id: fb947fe1749b3ecf13ad95394507dc54a8f6c213