SetDashboardInput: Move to c.g.g.extensions.api.projects
Moving the class to c.g.g.extensions.api.projects makes its package
consistent with the DashboardInfo and DashboardSectionInfo classes.

This move does not affect the API since none of the classes/methods
in the public API refer to it. It's only used in the DashboardApiImpl
and the dashboard REST API implementations.

Change-Id: I9ec61eaa89f974a872da3de20be8bc1fef40f581

Former-commit-id: e8754ecf093d923d85c36fa7965f3735dacfd15d