Set up Eclipse to work with Java 9 and later
To create .classpath with java 9, run:

  $ tools/eclipse/project.py --java 9

Start Eclipse, and follow the instructions provided in dev-eclipse.txt to
switch to Java 9 or later.

To run gerrit from Eclipse, the system classpath is cast to
URLClassLoader to retrieve the needed JARs. In Java 9 system classpath
cannot be cast to URLClassLoader. To rectify, extract the classpath from
Bazel.

Change-Id: Ie341cd4dfee5520f689fc50f34aa5395b74b5e7d

Former-commit-id: 8aae31c376ce1868dfbe28d1ef1a927e66e3a20d