Allow to assign "Delete Own Changes" permission to Change Owners
It is only possible for a user to delete their own change if they are
a member of a group that is assigned the Delete Own Changes permission.

This is counter-intuitive as it requires to either create a specific
group, or assign the permission to the Registered Users group.

Extend the permission checks to also allow assigning the permission
to the Change Owners group.

Change-Id: I13fa1bb1f4d0593d73cdb60d0fb4cb263de6c97c
Signed-off-by: Jacek Centkowski <jcentkowski@collab.net>
Signed-off-by: David Pursehouse <dpursehouse@collab.net>

Former-commit-id: 801bc2f8a4664a52af69651d03e8b6031cb4de46