Do not throw NotImplementedException from PublishChangeEdit and RebaseChangeEdit
Trying to list/parse members of the PublishChangeEdit and
RebaseChangeEdit REST collections returns '501 Not Implemented', but
usually we return '404 Not Found' for non-existing views. Make the
behaviour of PublishChangeEdit and RebaseChangeEdit consistent and throw
a ResourceNotFoundException instead.

Returning '501 Not Implemented' is bad since all 5xx responses are
counted into error statistics and failing when a non-implemented view is
invoked is not an error, but working as intended.

Change-Id: I6fcd80905cc280b434d75a5a6e6c312aa34b235a
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 62912f79def7684ff8e7a33fb2d26422cfe8ff21