Add --generate-http-password to ssh set-account
To bring the ssh command more in-line with the REST API for a user, it
is now possible to generate a new http-password which does not require
global admin (unlike direct set).

This allows ordinary users to set a new http password via ssh when they
cannot log in to the web ui (e.g. due to being a service account.)

The next commit will change the access restriction so that setting
your own account details does not require MODIFY_ACCOUNT. The REST API
does not need this, and granting it to someone allows them to set *any*
account details, e.g. even deactivating admin accounts.

Change-Id: Ic6102da9e483d07175cdc772930f25608dce12d4

Former-commit-id: efa38be397c174c444340371785fcf081515608d