Preserve author of submodule commits for super repo commit
If a super-repo is subscribed to 2 submodules and both submodules get
updated together by submitting a topic, a single commit in the
super-repo is created that updates both submodules at once. If the
commits in the submodules had the same author it was intended to
preserve this author for the super-repo commit. However when the authors
of the submodule commits were compared the author timestamp was taken
into account and as result the authors were (almost) never equal. Due to
this we were falling back to using the Gerrit server identity as author
for the commit in the super-repo. This is now fixed by ignoring the
author timestamp when comparing the authors.

Change-Id: I3130d29cb50abaf82e808523519d8743227e09d5
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 0fdd85fd3cc71777214c4b0896724c9b1c4f21c6