Always use new author timestamp for super repo commit
For commits in the super repo we preserve the author of the submodule
commits if all submodule commits have the same author. So far we also
used the same author timestamp. However if several submodules are
updated at once the submodule commits can have different author
timestamp, and we just used one of them for the super-repo commit. This
can be wrong as this way the author timestamp of the super-repo commit
might be earlier than some for some of the submodule commits. It's
better to always use a new author timestamp for the super-repo commit
(actual time of when the super-repo commit is created).

Change-Id: I74814095263530d1eaf6a4b332fc6a422929452b
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 08687ff3b8c796b2da27da41e2d3068849402c16