Dashboard: strip the Invalid-Prolog-Rules-Label-Name prefix from labels
This prefix is added by Gerrit (backend) to indicate the prolog rules
generate invalid label names, but causes issues because of the double
dash. The proposed fix is to strip this prefix on the dashboard, leading
to cleaner names - while not defeating the purpose (the ugly name is
still visible on the changes screen).

Change-Id: Ibfa37eb6231b48c2073948f26bd6b5ac30c15509

Former-commit-id: 4fde1339324e840a1aeae7c276e77f23ce294ed4