ElasticConfiguration: Make hosts private
Make the hosts member private and add a getter method. Move conversion
of the List to Array to this method.

In ElasticRestClientProvider, rather than keeping each used configuration
item as a member, keep the configuration and get each value from it as
needed. This will avoid having to add a new member for any new configs
that are added in future.

Change-Id: I711a53e90d59bb348a974872851856491755fb87

Former-commit-id: b6034ed80c06ae1939d754e72fec04d2225bd16f